move dat db to a .server file brudda
